Add optimistic uploads
Optimistically upload image files to hide the latency of image uploads.

This improves the experience of creating flashcards.
The downside is that we might have additional images that we don't want,
but this is not likely to occur frequently (images that we paste in the
input are usually the ones that we want to upload already).

In the future, I plan to add an interface that allows for managing the
images that we have locally and remotely, so the excess images shouldn't
be an issue.